@@ -11,7 +11,7 @@ repository = "https://github.com/BurntSushi/ripgrep/tree/master/crates/pcre2"
 readme = "README.md"
 keywords = ["regex", "grep", "pcre", "backreference", "look"]
 license = "Unlicense OR MIT"
-edition = "2018"
+edition = "2024"
 
 [dependencies]
 grep-matcher = { version = "0.1.7", path = "../matcher" }
