@@ -11,7 +11,7 @@ repository = "https://github.com/BurntSushi/ripgrep/tree/master/crates/searcher"
 readme = "README.md"
 keywords = ["regex", "grep", "egrep", "search", "pattern"]
 license = "Unlicense OR MIT"
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 bstr = { version = "1.6.2", default-features = false, features = ["std"] }
